Bug/dave 235 zeitraum ueber fuenf jahre (#109)
* util um zu prüfen ob der zeitraum größer 5 Jahre ist

* weiter prüfungen ob zeitraum größer 5 jahre
